+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMHQ6ETF10TW8Z8NG6TZ2GJP"
+        "01KMHXMHH7PAYQGTPHSNYYYNRM"
     )
     r"""The unique identifier of the data source"""