+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMHQ6ETE2RE3JYG1E1BX41DZ"
+        "01KMHXMHH6WM3XTZDYEB45A8C9"
     )
     r"""The unique identifier of the data source"""