+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMHQ6ETHKDACP6RBXVMFTNXD"
+        "01KMHXMHH9GNWY4SPRNEZ83J04"
     )
     r"""The unique identifier of the data source"""